Fixed the wrong story file
